requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
click==8.1.8
confluent-kafka==2.8.2
greenlet==3.1.1
h11==0.14.0
idna==3.10
kafka-python==2.1.2
psycopg2-binary==2.9.10
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
typing_extensions==4.12.2
uvicorn==0.34.0
